Separated ViewController and ViewModel in MovieInfoViewController file.
